Fix "Invoke-RestMethod:" error showing up when title is a program
The script first checks if what you have input is a series, and if it is, it proceeds without any complaints.

But if what you have input is a program, then you are met with something like this:

Invoke-RestMethod: C:\Users\Admin\Downloads\nrk-dl\nrk-dl.ps1:214
Line |
214 | … eries_req = Invoke-RestMethod -Uri "https://psapi.nrk.no/tv/catalog/s …
| ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
| "AAAA00000000 is not an identifier for a series, but an identifier for a program."

It will still download, but seeing as this might be confusing to new users, I have added try-catch blocks to handle this silently.